# Optimized label regex matcher with literal prefix and/or suffix