--- a/pkg/labels/labels_test.go
+++ b/pkg/labels/labels_test.go
@@ -368,7 +368,21 @@ func TestLabels_Equal(t *testing.T) {
 	}
 }
 
-func TestLabels_FromStringsPanic(t *testing.T) {
+func TestLabels_FromStrings(t *testing.T) {
+	labels := FromStrings("aaa", "111", "bbb", "222")
+	expected := Labels{
+		{
+			Name:  "aaa",
+			Value: "111",
+		},
+		{
+			Name:  "bbb",
+			Value: "222",
+		},
+	}
+
+	testutil.Equals(t, expected, labels, "unexpected labelset")
+
 	defer func() { recover() }()
 	FromStrings("aaa", "111", "bbb")