--- a/pkg/labels/regexp.go
+++ b/pkg/labels/regexp.go
@@ -1,0 +1,80 @@
+package labels
+
+import (
+	"regexp"
+	"regexp/syntax"
+	"strings"
+)
+
+type FastRegexMatcher struct {
+	re     *regexp.Regexp
+	prefix string
+	suffix string
+}
+
+func NewFastRegexMatcher(v string) (*FastRegexMatcher, error) {
+	re, err := regexp.Compile("^(?:" + v + ")$")
+	if err != nil {
+		return nil, err
+	}
+
+	parsed, err := syntax.Parse(v, syntax.Perl)
+	if err != nil {
+		return nil, err
+	}
+
+	m := &FastRegexMatcher{
+		re: re,
+	}
+
+	if parsed.Op == syntax.OpConcat {
+		m.prefix, m.suffix = optimizeConcatRegex(parsed)
+	}
+
+	return m, nil
+}
+
+func (m *FastRegexMatcher) MatchString(s string) bool {
+	if m.prefix != "" && !strings.HasPrefix(s, m.prefix) {
+		return false
+	}
+	if m.suffix != "" && !strings.HasSuffix(s, m.suffix) {
+		return false
+	}
+	return m.re.MatchString(s)
+}
+
+func (m *FastRegexMatcher) GetRegexString() string {
+	return m.re.String()
+}
+
+// optimizeConcatRegex returns literal prefix/suffix text that can be safely
+// checked against the label value before running the regexp matcher.
+func optimizeConcatRegex(r *syntax.Regexp) (prefix, suffix string) {
+	sub := r.Sub
+
+	// We can safely remove begin and end text matchers respectively
+	// at the beginning and end of the regexp.
+	if len(sub) > 0 && sub[0].Op == syntax.OpBeginText {
+		sub = sub[1:]
+	}
+	if len(sub) > 0 && sub[len(sub)-1].Op == syntax.OpEndText {
+		sub = sub[:len(sub)-1]
+	}
+
+	if len(sub) == 0 {
+		return
+	}
+
+	// Given Prometheus regex matchers are always anchored to the begin/end
+	// of the text, if the first/last operations are literals, we can safely
+	// treat them as prefix/suffix.
+	if sub[0].Op == syntax.OpLiteral {
+		prefix = string(sub[0].Rune)
+	}
+	if last := len(sub) - 1; sub[last].Op == syntax.OpLiteral {
+		suffix = string(sub[last].Rune)
+	}
+
+	return
+}