@codesome 2020-06-25
> Might it make sense to push this upstream, rather than us slowly writing our own regex engine?

I feel that would slow down the development, and we would only optimize those regexes that are of interest to PromQL usage, so it might make more sense here. Anything which looks very general we could put it upstream and remove it from here.