@dgl 2020-06-25
> That's a good point. I was as well, especially because the regex engine has an optimization for prefixes

I think the optimisation only applies for unanchored patterns, you can see if you use [LiteralPrefix](https://golang.org/pkg/regexp/#Regexp.LiteralPrefix) as a proxy for the
prefix optimisation that it doesn't work for the patterns Prometheus will generate (https://play.golang.org/p/NPfMJTqwJaj). Definitely worth asking the Go people.