@codesome 2020-06-26
I think this PR is good to go. Discussions in Go can follow this. I will merge this in ~2h if no concerns.